+++ java8/src/main/java/io/cucumber/java8/Java8DefaultDataTableCellTransformerDefinition.java
@@ -3,8 +3,6 @@ package io.cucumber.java8;
 import io.cucumber.core.backend.DefaultDataTableCellTransformerDefinition;
 import io.cucumber.datatable.TableCellByTypeTransformer;
 
-import java.lang.reflect.Type;
-
 class Java8DefaultDataTableCellTransformerDefinition extends AbstractDatatableElementTransformerDefinition
         implements DefaultDataTableCellTransformerDefinition {
 
@@ -14,13 +12,9 @@ class Java8DefaultDataTableCellTransformerDefinition extends AbstractDatatableEl
 
     @Override
     public TableCellByTypeTransformer tableCellByTypeTransformer() {
-        return (fromValue, toValueType) -> execute(
+        return (fromValue, toValueType) -> invokeMethod(
             replaceEmptyPatternsWithEmptyString(fromValue),
             toValueType);
-    }
-
-    private Object execute(String fromValue, Type toValueType) {
-        return Invoker.invoke(this, body, method, fromValue, toValueType);
     }
 
 }