+++ java8/src/main/java/io/cucumber/java8/Java8DefaultDataTableEntryTransformerDefinition.java
@@ -3,9 +3,6 @@ package io.cucumber.java8;
 import io.cucumber.core.backend.DefaultDataTableEntryTransformerDefinition;
 import io.cucumber.datatable.TableEntryByTypeTransformer;
 
-import java.lang.reflect.Type;
-import java.util.Map;
-
 class Java8DefaultDataTableEntryTransformerDefinition extends AbstractDatatableElementTransformerDefinition
         implements DefaultDataTableEntryTransformerDefinition {
 
@@ -20,13 +17,8 @@ class Java8DefaultDataTableEntryTransformerDefinition extends AbstractDatatableE
 
     @Override
     public TableEntryByTypeTransformer tableEntryByTypeTransformer() {
-        return (fromValue, toValueType, tableCellByTypeTransformer) -> execute(
+        return (fromValue, toValueType, tableCellByTypeTransformer) -> invokeMethod(
             replaceEmptyPatternsWithEmptyString(fromValue),
             toValueType);
     }
-
-    private Object execute(Map<String, String> fromValue, Type toValueType) {
-        return Invoker.invoke(this, body, method, fromValue, toValueType);
-    }
-
 }