+++ java8/src/main/java/io/cucumber/java8/Java8DocStringTypeDefinition.java
@@ -20,11 +20,7 @@ final class Java8DocStringTypeDefinition extends AbstractGlueDefinition implemen
         this.docStringType = new DocStringType(
             returnType,
             contentType,
-            this::execute);
-    }
-
-    private Object execute(String content) {
-        return Invoker.invoke(this, body, method, content);
+            this::invokeMethod);
     }
 
     @Override